sender/plain: stream MIME generation to SMTP host
Adds the streaming MIME generation from the TLS sender to the plain-text SMTP
sender.

Fixes a bug where the hostname and port was used as the hostname - this caused
errors to be returned when working with local SMTP hosts that require AUTH, and
do not support STARTTLS.
