+++ integration_test.go
@@ -88,7 +88,7 @@ func TestIntegration_TLS(t *testing.T) {
 				m.From("dom@itsallbroken.com")
 				m.FromName("Dom")
 				m.To("to@example.org")
-				m.Subject("Attachment test")
+				m.Subject("TLS Attachment test")
 				m.ReplyTo("replies@example.org")
 				m.HTML().Set("Attachment MF5: " + hashString)
 				m.Attach("test.bin", bytes.NewReader(data))
@@ -142,7 +142,7 @@ func TestIntegration_PlainText(t *testing.T) {
 				m.FromName("From Example")
 				m.To("to@example.org")
 				m.Bcc("bcc1@example.org", "bcc2@example.org")
-				m.Subject("Test subject")
+				m.Subject("PLAIN - Test subject")
 				m.ReplyTo("replies@example.org")
 				m.HTML().Set("HTML part: this is just a test.")
 				m.Plain().Set("Plain text part: this is also just a test.")
@@ -160,7 +160,7 @@ func TestIntegration_PlainText(t *testing.T) {
 				m.FromName("From Example")
 				m.To("to@example.org")
 				m.Bcc("bcc1@example.org", "bcc2@example.org")
-				m.Subject("TLS test")
+				m.Subject("PLAIN - TLS test")
 				m.ReplyTo("replies@example.org")
 				m.HTML().Set("HTML part: this is just a test.")
 				m.Plain().Set("Plain text part: this is also just a test.")
@@ -181,7 +181,7 @@ func TestIntegration_PlainText(t *testing.T) {
 				m.From("dom@itsallbroken.com")
 				m.FromName("Dom")
 				m.To("to@example.org")
-				m.Subject("Attachment test")
+				m.Subject("PLAIN - Attachment test")
 				m.ReplyTo("replies@example.org")
 				m.HTML().Set("Attachment MF5: " + hashString)
 				m.Attach("test.bin", bytes.NewReader(data))
